Removed and Updated GET_PUBLIC_LEARNING_OBJECTS and GET_PUBLIC_LEARNING_OBJECTS_WITH_FILTER to point to search.service.ts.
