+++ src/app/core/learning-object-module/search/search.service.ts
@@ -1,9 +1,96 @@
 import { Injectable } from '@angular/core';
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
+import { LearningObject } from '@entity';
+import { catchError } from 'rxjs/operators';
+import { throwError } from 'rxjs';
+import { Query } from 'app/interfaces/query';
+import * as querystring from 'querystring';
+import { SEARCH_ROUTES } from 'app/core/learning-object-module/search/search.routes';
+
 
 @Injectable({
   providedIn: 'root'
 })
 export class SearchService {
 
-  constructor() { }
+  constructor(private http: HttpClient) { }
+
+  /**
+   * Fetches Array of Learning Objects
+   *
+   * @returns {Promise<LearningObject[]>}
+   * @memberof LearningObjectService
+   */
+  getLearningObjects(
+    query?: Query,
+  ): Promise<{ learningObjects: LearningObject[]; total: number }> {
+    let route = '';
+    if (query) {
+      const queryClone = Object.assign({}, query);
+      if (
+        queryClone.standardOutcomes &&
+        queryClone.standardOutcomes.length &&
+        typeof queryClone.standardOutcomes[0] !== 'string'
+      ) {
+        queryClone.standardOutcomes = (
+          queryClone.standardOutcomes as string[]
+        ).map((o) => o['id']);
+      }
+      const queryString = new URLSearchParams(queryClone).toString();
+      route =
+        SEARCH_ROUTES.GET_PUBLIC_LEARNING_OBJECTS_WITH_FILTER(queryString);
+    } else {
+      route = SEARCH_ROUTES.GET_PUBLIC_LEARNING_OBJECTS;
+    }
+
+    return this.http
+      .get(route)
+      .pipe(catchError(this.handleError))
+      .toPromise()
+      .then((response: any) => {
+        const objects = response.objects;
+        return {
+          learningObjects: objects.map((object) => new LearningObject(object)),
+          total: response.total,
+        };
+      });
+  }
+
+    /**
+   * Fetches Array of Learning Objects that are not currently featured
+   *
+   * @returns {Promise<LearningObject[]>}
+   * @memberof LearningObjectService
+   */
+    getNotFeaturedLearningObjects(
+      query?: Query,
+    ): Promise<{ learningObjects: LearningObject[]; total: number }> {
+      let route = '';
+      if (query) {
+        const queryClone = Object.assign({}, query);
+        const queryString = querystring.stringify(queryClone);
+        route =
+          SEARCH_ROUTES.GET_PUBLIC_LEARNING_OBJECTS_WITH_FILTER(queryString);
+      } else {
+        route = SEARCH_ROUTES.GET_PUBLIC_LEARNING_OBJECTS;
+      }
+  
+      return this.http
+        .get(route)
+        .pipe(catchError(this.handleError))
+        .toPromise()
+        .then((response: any) => {
+          return { learningObjects: response.objects, total: response.total };
+        });
+    }
+
+    private handleError(error: HttpErrorResponse) {
+      if (error.error instanceof ErrorEvent) {
+        // Client-side or network returned error
+        return throwError(error.error.message);
+      } else {
+        // API returned error
+        return throwError(error);
+      }
+    }
 }
