+++ src/app/cube/learning-object.service.ts
@@ -7,10 +7,6 @@ import {
   LEARNING_OBJECT_ROUTES,
   LEGACY_USER_ROUTES
 } from '../core/learning-object-module/learning-object/learning-object.routes';
-import { Query } from '../interfaces/query';
-
-import { SEARCH_ROUTES } from 'app/core/learning-object-module/search/search.routes';
-import * as querystring from 'querystring';
 import { REVISION_ROUTES } from '../core/learning-object-module/revisions/revisions.routes';
 import { USER_ROUTE } from '../core/user-module/user.routes';
 
@@ -40,46 +36,7 @@ export class LearningObjectService {
   openLearningObject(url: string) {
     window.open(url);
   }
-  /**
-   * Fetches Array of Learning Objects
-   *
-   * @returns {Promise<LearningObject[]>}
-   * @memberof LearningObjectService
-   */
-  getLearningObjects(
-    query?: Query,
-  ): Promise<{ learningObjects: LearningObject[]; total: number }> {
-    let route = '';
-    if (query) {
-      const queryClone = Object.assign({}, query);
-      if (
-        queryClone.standardOutcomes &&
-        queryClone.standardOutcomes.length &&
-        typeof queryClone.standardOutcomes[0] !== 'string'
-      ) {
-        queryClone.standardOutcomes = (
-          queryClone.standardOutcomes as string[]
-        ).map((o) => o['id']);
-      }
-      const queryString = new URLSearchParams(queryClone).toString();
-      route =
-        SEARCH_ROUTES.GET_PUBLIC_LEARNING_OBJECTS_WITH_FILTER(queryString);
-    } else {
-      route = SEARCH_ROUTES.GET_PUBLIC_LEARNING_OBJECTS;
-    }
-
-    return this.http
-      .get(route)
-      .pipe(catchError(this.handleError))
-      .toPromise()
-      .then((response: any) => {
-        const objects = response.objects;
-        return {
-          learningObjects: objects.map((object) => new LearningObject(object)),
-          total: response.total,
-        };
-      });
-  }
+  
 
   /**
    * Fetches LearningObject by cuid
